requirements.txt
streamlit
streamlit_chat
langchain
sentence_transformers
pinecone-client
torch
transformers
openai
python-dotenv
tiktoken
